Text is now drawn using float positions not ints

+++ src/main/java/coffeecatteam/theultimatetile/gfx/Text.java
@@ -7,32 +7,32 @@ import org.newdawn.slick.Graphics;
 
 public class Text {
 
-    public static void drawStringCenteredX(Graphics g, String text, int yPos, Color c, Font font) {
-        int width = getWidth(text, font);
-        drawString(g, text, TutEngine.getTutEngine().getWidth() / 2 - width / 2, yPos, false, c, font);
+    public static void drawStringCenteredX(Graphics g, String text, float yPos, Color c, Font font) {
+        float width = getWidth(text, font);
+        drawString(g, text, TutEngine.getTutEngine().getWidth() / 2f - width / 2, yPos, false, c, font);
     }
 
-    public static void drawStringCenteredY(Graphics g, String text, int xPos, Color c, Font font) {
-        int height = getHeight(text, font);
-        drawString(g, text, xPos, (TutEngine.getTutEngine().getHeight() / 2 - height / 2) + getAscent(font), false, c, font);
+    public static void drawStringCenteredY(Graphics g, String text, float xPos, Color c, Font font) {
+        float height = getHeight(text, font);
+        drawString(g, text, xPos, (TutEngine.getTutEngine().getHeight() / 2f - height / 2) + getAscent(font), false, c, font);
     }
 
     public static void drawStringCentered(Graphics g, String text, Color c, Font font) {
-        int width = getWidth(text, font), height = getHeight(text, font);
-        int x = TutEngine.getTutEngine().getWidth() / 2 - width / 2, y = (TutEngine.getTutEngine().getHeight() / 2 - height / 2) + getAscent(font);
+        float width = getWidth(text, font), height = getHeight(text, font);
+        float x = TutEngine.getTutEngine().getWidth() / 2f - width / 2, y = (TutEngine.getTutEngine().getHeight() / 2f - height / 2) + getAscent(font);
         drawString(g, text, x, y, false, c, font);
     }
 
-    public static void drawString(Graphics g, String text, int xPos, int yPos, boolean centered, Color c, Font font) {
+    public static void drawString(Graphics g, String text, float xPos, float yPos, boolean centered, Color c, Font font) {
         drawString(g, text, xPos, yPos, centered, centered, c, font);
     }
 
-    public static void drawString(Graphics g, String text, int xPos, int yPos, boolean centeredX, boolean centeredY, Color c, Font font) {
+    public static void drawString(Graphics g, String text, float xPos, float yPos, boolean centeredX, boolean centeredY, Color c, Font font) {
         g.setAntiAlias(false);
         g.setColor(c);
         g.setFont(font);
-        int x = xPos;
-        int y = yPos - getHeight(text, font);
+        float x = xPos;
+        float y = yPos - getHeight(text, font);
         if (centeredX)
             x = xPos - getWidth(text, font) / 2;
         if (centeredY)
@@ -40,15 +40,15 @@ public class Text {
         g.drawString(text, x, y);
     }
 
-    public static int getWidth(String text, Font font) {
+    public static float getWidth(String text, Font font) {
         return font.getWidth(text);
     }
 
-    public static int getHeight(String text, Font font) {
+    public static float getHeight(String text, Font font) {
         return font.getHeight(text);
     }
 
-    public static int getAscent(Font font) {
+    public static float getAscent(Font font) {
         return font.getLineHeight();
     }
 }
